% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\name{add_contours}
\alias{add_contours}
\title{Print contour lines onto the retina plot}
\usage{
add_contours(minitics, Mat, contour_breaks, xy)
}
\arguments{
  \item{minitics}{See fit_plot_azimuthal} \item{Mat}
  {See fit_plot_azimuthal} \item{contour_breaks}{See
  fit_plot_azimuthal}
}
\description{
Makes a set of contours to the retinaplot. Modified
code from http://stackoverflow.com/questions/10856882/r-
interpolated-polar-contour-plot was highly modified to meet
retinal plotting funtionality.
}

