% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_degrees}
\alias{add_degrees}
\title{Add Degrees}
\usage{
add_degrees(theta, degrees)
}
\arguments{
  \item{theta}{Angle in degrees}

  \item{degrees}{Number of degrees to add to theta}
}
\value{
newtheta Adjusted between -180 and 180 degrees.
}
\description{
Add Degrees
}
\examples{
add_degrees(30, 180) #you should get -150
}
\author{
Brian Cohn \email{brian_cohn14@pitzer.edu}, Lars Schmitz
}
\seealso{
Other trig_fns: \code{\link{cartesian_rotation}}
}

