% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\name{add_legend}
\alias{add_legend}
\title{add a basic legend}
\usage{
add_legend(col, zlim)
}
\arguments{
  \item{col}{Color vector} \item{zlim}{limits of the
  densities}
}
\description{
plot a basic legend
}
\author{
Brian Cohn
}

