% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{assemble_markup_file}
\alias{assemble_markup_file}
\title{Assemble markup file}
\usage{
assemble_markup_file(eye_left_or_right, path_to_retina_data_folder,
  dorsal_outline_index = NA, nasal_outline_index = NA)
}
\arguments{
\item{eye_left_or_right}{either 'left' or 'right'}

\item{path_to_retina_data_folder}{The path where the markup.csv file will be saved.}

\item{dorsal_outline_index}{The index of the point that points toward dorsal}

\item{nasal_outline_index}{The index of the opint that points toward nasal.}

\item{markup_coordinates_dataframe}{markup metaparameters in retistruct format}
}
\description{
Creates a markup.csv file. One of the dorsal_outline_index or nasal_outline_index must be filled in. Not both.
}
\author{
Brian Cohn
}
