% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{assemble_point_coordinates_file}
\alias{assemble_point_coordinates_file}
\title{Assemble outline P.csv XY coordinates file}
\usage{
assemble_point_coordinates_file(outline_coordinates, path_to_retina_data_folder)
}
\arguments{
\item{outline_coordinates}{the dataframe of 2 columns with XY pixel points}

\item{path_to_retina_data_folder}{The path where the P.csv file will be saved.}

\item{outline_coordinates}{outline coords in retistruct format}
}
\description{
Creates a P.csv file
}
\author{
Brian Cohn
}
