% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic-s3.R
\name{bitwAnd}
\alias{bitwAnd}
\alias{bitwAnd.default}
\alias{bitwAnd.Image}
\alias{bitwOr}
\alias{bitwOr.default}
\alias{bitwOr.Image}
\alias{bitwXor}
\alias{bitwXor.default}
\alias{bitwXor.Image}
\alias{bitwNot}
\alias{bitwNot.default}
\alias{bitwNot.Image}
\title{Bitwise operations for Image objects}
\usage{
bitwAnd(x, y, ...)

\method{bitwAnd}{default}(x, y, ...)

\method{bitwAnd}{Image}(x, y, ...)

bitwOr(x, y, ...)

\method{bitwOr}{default}(x, y, ...)

\method{bitwOr}{Image}(x, y, ...)

bitwXor(x, y, ...)

\method{bitwXor}{default}(x, y, ...)

\method{bitwXor}{Image}(x, y, ...)

bitwNot(x, ...)

\method{bitwNot}{default}(x, ...)

\method{bitwNot}{Image}(x, ...)
}
\arguments{
\item{x}{An \code{Image} object, or an integer vector for the default
methods.}

\item{y}{An \code{Image} object, scalar, or integer vector
(\code{bitwAnd}, \code{bitwOr}, \code{bitwXor} only).}

\item{...}{Additional arguments (ignored for \code{Image} objects).}
}
\value{
For \code{Image} inputs, a new \code{Image} with the element-wise
  bitwise result. For other inputs, delegates to the corresponding
  \code{\link[base]{bitwAnd}} family function.
}
\description{
S3 generics extending \code{\link[base]{bitwAnd}}, \code{\link[base]{bitwOr}},
\code{\link[base]{bitwXor}}, and \code{\link[base]{bitwNot}} to support
\code{Image} objects. Non-Image inputs delegate to the corresponding
\code{base} function.
}
\seealso{
\code{\link[base]{bitwAnd}}
}
