% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\name{cartesian_rotation}
\alias{cartesian_rotation}
\title{Counterclockwise rotation about the origin}
\usage{
cartesian_rotation(x, y, theta)
}
\arguments{
\item{x}{vector of x coordinates (int/float)}

\item{y}{vector of y coordinates (int/float)}

\item{theta}{(float|int) angle in degrees.}
}
\value{
xy Data.Frame with $x and $y rotated coordinates
}
\examples{
x<- c(1,2,3,4,5)
y<- c(3,3,3,3,3)
theta <- 30
cartesian_rotation(x,y,theta)
}
\seealso{
Other trig_fns: \code{\link{add_degrees}}
}
\author{
Brian Cohn \email{brian_cohn14@pitzer.edu}, Lars Schmitz
}
