% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retistruct_processing_functions.R
\name{compose_tear_triplets_dataframe}
\alias{compose_tear_triplets_dataframe}
\title{Set up tear matrix}
\usage{
compose_tear_triplets_dataframe(list_of_tear_triplets, outline_object)
}
\arguments{
\item{list_of_tear_triplets}{list of 3-element vectors}

\item{outline_object}{See ?retistruct::labelTearPoints for documentation}
}
\value{
mat matrix of tear vals, ready for T.csv saving
}
\description{
takes in a list of tear triplets (numeric vectors of 3 elements) and outputs the dataframe, ready for retistruct processing
}
\author{
Brian Cohn \email{brian.cohn@usc.edu}
}
