% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{compute_longitude_label_location}
\alias{compute_longitude_label_location}
\title{Compute Longitude Label Location}
\usage{
compute_longitude_label_location(axis.rads, outer_radius)
}
\arguments{
\item{axis.rads}{axis.rads}

\item{outer_radius}{numeric value for radius limit}
}
\value{
label_locations the computed location for a label
}
\description{
Find the location to put the label around the circle at each of the lines
}
\author{
Brian Cohn
}
