% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compute_rotation_matrix}
\alias{compute_rotation_matrix}
\title{Compute rotation matrix}
\usage{
compute_rotation_matrix(retina_list, spatial_res = 16, theta_interval = 10)
}
\arguments{
  \item{retina_list}{list of retinal objects. for example,
  list(Pmol_753=Pmol_753, Pmol_752=Pmol_752,
  Ntae_381=Ntae_381)}

  \item{spatial_res}{width of the evaluation grid that's
  used for every combination.}

  \item{theta_interval}{This value (an integer in degrees)
  will generate (360/theta_interval) rotations to
  evaluate.}
}
\value{
rotation_mat Matrix where columns are fixed, rows are
rotated. Column 1 is the first element in the retina_list
}
\description{
Fixes each retina in the inputted list of retina-objects,
generating the optimal rotation (in degrees) with respect
to the other retinas.
}
\author{
Brian Cohn \email{brian_cohn14@pitzer.edu} and Lars Schmitz
}

