% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\name{coordinate_IJ}
\alias{coordinate_IJ}
\title{ImageJ Coordinate Conversion}
\usage{
coordinate_IJ(RET_count_data, maxX, maxY, minX, minY, deltaX, deltaY, inversion,
  ...)
}
\arguments{
\item{RET_count_data}{XYZ dataframe. X and Y are integer values starting at 1.}

\item{maxX}{The X value of the furthest sampling site at the far right of the picture.}

\item{maxY}{The Y value of the furthest sampling site at the top of the picture}

\item{minX}{The X value of the furthest sampling site at the far left of the picture}

\item{minY}{The Y value of the furthest sampling site at the bottom of the picture}

\item{deltaX}{Average change in pixels in the X axis between two sampling sites.}

\item{deltaY}{Average change in pixels in the Y axis between two sampling sites.}

\item{inversion}{Whether or not to invert}

\item{...}{arguments passed to or from other methods.}
}
\value{
RET_count_data Retinal count data in ImageJ coordinates. $x, $y and $z
}
\description{
It's important for the user to communicate the dimensions of the retinal wholemount to the program that's going to reconstruct the original shape. coordinate_IJ is useful for noting what the grid looks like for a retinal wholemount, so the package can decide where to place the sampling locations with respect to the outline created in ImageJ.
}
\note{
Need to remove all imageJ calibration in an Epic
}
\author{
Brian Cohn \email{brian.cohn@usc.edu}, Lars Schmitz
}
