% Generated by roxygen2 (4.0.0): do not edit by hand
\name{count_to_rho}
\alias{count_to_rho}
\title{Convert count to Density}
\usage{
count_to_rho(count, height, width)
}
\arguments{
  \item{count}{Number of retinal cells counted within the
  counting frame.}

  \item{height}{Counting frame width in microns}

  \item{width}{Counting frame width in microns}
}
\value{
rho Density in units of retinal ganglion cells per square
millimeter
}
\description{
Uses cell count to compute density within the counting
frame. Units are set to microns
}
\examples{
count_to_rho(10,25,25)
#16,000 cells/mm^2
}
\author{
Brian Cohn \email{brian_cohn14@pitzer.edu}
}

