% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{define_color_breaks_based_on_source}
\alias{define_color_breaks_based_on_source}
\title{Define Zlim by the requested color breaks source}
\usage{
define_color_breaks_based_on_source(col_breaks_source, z, Mat)
}
\arguments{
\item{col_breaks_source}{A 2 element vector with max and min}

\item{z}{the response values from the input data (the retinal densities)}

\item{Mat}{The predicted retinal densities across the xy space}
}
\description{
Set color breaks (zlim) based on requested source
}
\author{
Brian Cohn
}
