% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\name{degrees}
\alias{degrees}
\title{Conversion from Radians to Degrees}
\usage{
degrees(radian_coords)
}
\arguments{
\item{radian_coords}{A data.frame with two columns- first is phi, second is lambda}
}
\value{
data.frame Phi and lambda in a data.frame.
}
\author{
Brian Cohn \email{brian.cohn@usc.edu}, Lars Schmitz
}
