% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\name{draw_axial_spokes}
\alias{draw_axial_spokes}
\title{draw axial spokes}
\usage{
draw_axial_spokes(endpoints, color_hex = "#66666650")
}
\arguments{
  \item{endpoints}{a 4 element numeric vector descrbing the xy
  and x'y' for the line segment.} \item{color_hex}{hex string}
}
\description{
put the radial lines on the plot
}
\author{
Brian Cohn
}

