% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{draw_line_segments}
\alias{draw_line_segments}
\title{draw line segments}
\usage{
draw_line_segments(endpoints, color_hex = "#66666650")
}
\arguments{
\item{endpoints}{a 4 element numeric vector descrbing the xy and x'y' for the line segment.}

\item{color_hex}{hex string}
}
\description{
put the radial lines on the plot
}
\author{
Brian Cohn
}
