% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retistruct_processing_functions.R
\name{dss_retistruct_processing}
\alias{dss_retistruct_processing}
\title{Retistruct Wrapper}
\usage{
dss_retistruct_processing(path)
}
\arguments{
\item{path}{string path to the folder with retinal data.}
}
\value{
rad Retinal data in radian units.
}
\description{
Reads in the folder, reconstructs, and returns the retinal object.
}
\references{
Sterratt et. al. 2013
}
\author{
Brian Cohn \email{brian.cohn@usc.edu}
}
