% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{extract_xy_and_falciform}
\alias{extract_xy_and_falciform}
\title{Extract xy points and falciform}
\usage{
extract_xy_and_falciform(datapoints_from_retistruct, number_of_sampling_sites)
}
\arguments{
\item{datapoints_from_retistruct}{data structure that comes from the datapoints.csv, which is automatically generated by Retistruct.}

\item{number_of_sampling_sites}{integer}
}
\value{
L list of pixel xy positions, and falciform coordinates in [1] and [2], respectively
}
\description{
Extracts xy and falciform coordinates from datapoints collected from retistruct. Falc and Datapoints are stacked upon one another
}
\author{
Brian Cohn
}
