% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\name{fit_error_histogram}
\alias{fit_error_histogram}
\title{Single Tps Fit Error Plot}
\usage{
fit_error_histogram(x, main = NULL, ...)
}
\arguments{
\item{x}{the fit object}

\item{main}{title of the plot, NULL by default}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Error histogram in base R
}
\description{
This function helps the user answer the scientific question: What is the fit performance of the Thin Plate Spline that is used to smooth the surface of the retinal density map? It is also useful for identifying systematic bias, bimodal distributions, or any other issues that could interfere with statistical rigor.
}
\author{
Brian Cohn \email{brian.cohn@usc.edu}, Lars Schmitz
}
