% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{fit_plot_azimuthal}
\alias{fit_plot_azimuthal}
\title{Polar Interpolation}
\usage{
fit_plot_azimuthal(x, y, z, contours = TRUE, legend = TRUE, axes = TRUE,
  should_plot_points = TRUE, extrapolate = TRUE, col_breaks_source = 2,
  col_levels = 50, col = rev(colorRampPalette(brewer.pal(11,
  "PuOr"))(col_levels)), contour_breaks_source = 1,
  contour_levels = col_levels + 1, outer.radius = pi/2,
  circle.rads = pretty_list_not_including_max(0, outer.radius),
  spatial_res = 1000, single_point_overlay = 0, lambda = 0.001,
  xyrelief = 1, tmp_input = NULL, plot_suppress = FALSE,
  compute_error = FALSE, falciform_coords = NULL, falc2 = NA,
  polynomial_m = NULL, ...)
}
\arguments{
\item{x, y, z}{cartesian input in azimuthal format}

\item{contours}{whether to plot contours.}

\item{legend}{Color legend with tick marks}

\item{axes}{Radial axes}

\item{extrapolate}{By default TRUE, will make plot a circle.}

\item{col_breaks_source}{2 element vector with max and min}

\item{col_levels}{number of color levels}

\item{col}{colors to plot}

\item{contour_breaks_source}{1 if data, 2 if calculated surface data}

\item{contour_levels}{number of contour levels}

\item{outer.radius}{size of plot}

\item{circle.rads}{radius lines}

\item{spatial_res}{Used to define a spatial_res by spatial_res plotting resolution.}

\item{lambda}{lambda value for thin plate spline interpolation}

\item{xyrelief}{scaling factor for interpolation matrix.}

\item{tmp_input}{tmp_input}

\item{plot_suppress}{by default FALSE}

\item{compute_error}{whether to use fields::predictSE}

\item{falciform_coords}{vertices in xy format of the falciform process}

\item{falc2}{a second falficorm coordinate file}

\item{polynomial_m}{A polynomial function of degree (m-1) will be included in the model as the drift (or spatial trend) component. Default is the value such that 2m-d is greater than zero where d is the dimension of x.}

\item{...}{passed arguments}

\item{points}{whether to plot individual datapoints as X's}
}
\description{
This function will make a plot. Code from http://stackoverflow.com/questions/10856882/r-interpolated-polar-contour-plot was highly modified to meet retinal plotting funtionality.
}
