% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\name{fit_plots}
\alias{fit_plots}
\title{Retinal Krig Fit Plots}
\usage{
fit_plots(x, digits = 4, which = 1:4, ...)
}
\arguments{
\item{x}{the fit object}

\item{digits}{set to 4 arbitrarily}

\item{which}{set arbitrarily to 1:4}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Histograms and Error scatterplots in base R
}
\references{
Fields Package
}
\author{
Brian Cohn \email{brian_cohn14@pitzer.edu}, Lars Schmitz
}
