% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retistruct_processing_functions.R
\name{generate_outline_with_tears}
\alias{generate_outline_with_tears}
\title{Generate outline with tears}
\usage{
generate_outline_with_tears(outline_coordinates, list_of_tear_triplets,
  path_to_retina_data_folder)
}
\arguments{
\item{outline_coordinates}{Outline coordinates XY str generated from tear_markup_plot function}

\item{list_of_tear_triplets}{list of 3-element vectors}

\item{path_to_retina_data_folder}{path to working directory}
}
\value{
outline_with_tears AnnotatedOutline. See ?retistruct:::AnnotatedOutline
}
\description{
Creates a valid AnnotatedOutline, ready for retistruct
}
\author{
Brian Cohn \email{brian.cohn@usc.edu}
}
