% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\name{get_path_to_test_retina_folder}
\alias{get_path_to_test_retina_folder}
\title{get_path_to_test_retina_folder
useful for access of the test retina datasets'}
\usage{
get_path_to_test_retina_folder(string_identifier)
}
\arguments{
\item{string_identifier}{character e.g. "40oik5" that represents the inst/extdata identifier folder name.}
}
\value{
filepath to the test_retina folder matching the string identifier variable.
}
\description{
get_path_to_test_retina_folder
useful for access of the test retina datasets'
}
