% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/retina-package.r
\name{import_xyz}
\alias{import_xyz}
\title{import_xyz}
\usage{
import_xyz(path)
}
\arguments{
  \item{path}{Path to a folder containing a three column
  comma separated value database. File must be named
  \code{xyz.csv}}
}
\value{
dataframe with X Y and Z columns with NA's. Each row is a
density measurement at a given XY location in ImageJ
coordinates.
}
\description{
Grabs the XYZ from the path
}
\author{
Brian Cohn \email{brian_cohn14@pitzer.edu}, Lars Schmitz
}
\seealso{
Other munge: \code{\link{ssite_merge}}
}

