% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\name{init_square_mat_plot}
\alias{init_square_mat_plot}
\title{Initiate the Square Matrix plot to prepare for polar plotting}
\usage{
init_square_mat_plot(Mat, zlim, minitics, col)
}
\arguments{
\item{Mat}{See fit_plot_azimuthal}

\item{zlim}{See fit_plot_azimuthal}

\item{minitics}{See fit_plot_azimuthal}

\item{col}{See fit_plot_azimuthal}
}
\description{
instantiates the square plotting area Modified code from http://stackoverflow.com/questions/10856882/r-interpolated-polar-contour-plot was highly modified to meet retinal plotting funtionality.
}
