% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\name{interpolate_input_data}
\alias{interpolate_input_data}
\title{Interpolate Input Data with Thin Plate Spline}
\usage{
interpolate_input_data(minitics, x, y, z, lambda, polynomial_m, extrapolate)
}
\arguments{
  \item{minitics}{values referring to the circle} \item{x}
  {input variable 1} \item{y}{input variable 2} \item{z}
  {response variable} \item{lambda}{TPS parameter}
  \item{polynomial_m}{TPS parameter} \item{extrapolate}{true/
  false whether we should interpolate past the points, all the
  way to the eye equator}
}
\description{
Interpolation
}
\author{
Brian Cohn
}

