% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\name{load_function_and_run}
\alias{load_function_and_run}
\title{load_function_and_run
useful for access of the test retina datasets'}
\usage{
load_function_and_run(string_identifier)
}
\arguments{
\item{string_identifier}{character e.g. "40oik5" that represents the inst/extdata identifier folder name.}
}
\description{
load_function_and_run
useful for access of the test retina datasets'
}
