% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels.R
\name{merge_channels}
\alias{merge_channels}
\title{Merge a list of single-channel images into a multi-channel image}
\usage{
merge_channels(channels)
}
\arguments{
\item{channels}{A named list of single-channel \code{Image} objects with
equal dimensions and depth. Names are used to infer the output colorspace.}
}
\value{
A new \code{Image}. Colorspace is inferred from
  \code{names(channels)}; if unrecognised, colorspace is set to
  \code{"UNKNOWN"} with a warning.
}
\description{
Merge a list of single-channel images into a multi-channel image
}
\examples{
\donttest{
img <- Image$new(system.file("img", "flower.jpg", package = "Retina"))
channels <- split_channels(img)
reconstructed <- merge_channels(channels)
reconstructed$plot()
}
}
