% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology.R
\name{morph_blackhat}
\alias{morph_blackhat}
\title{Apply black-hat morphological transform to an image}
\usage{
morph_blackhat(img, ...)
}
\arguments{
\item{img}{An \code{Image} object.}

\item{...}{Passed to \code{img$morph("blackhat", ...)}.}
}
\value{
A new \code{Image}.
}
\description{
Apply black-hat morphological transform to an image
}
\examples{
\donttest{
img_path <- system.file("img", "flower.jpg", package = "Retina")
img <- Image$new(img_path)$convert_color("GRAY")
morph_blackhat(img)$plot()
}
}
