% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology.R
\name{morph_erode}
\alias{morph_erode}
\title{Erode an image}
\usage{
morph_erode(img, ...)
}
\arguments{
\item{img}{An \code{Image} object.}

\item{...}{Passed to \code{img$morph("erode", ...)}.}
}
\value{
A new \code{Image}.
}
\description{
Erode an image
}
\examples{
\donttest{
img_path <- system.file("img", "flower.jpg", package = "Retina")
img <- Image$new(img_path)$convert_color("GRAY")
morph_erode(img)$plot()
}
}
