% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology.R
\name{morph_gradient}
\alias{morph_gradient}
\title{Compute the morphological gradient of an image}
\usage{
morph_gradient(img, ...)
}
\arguments{
\item{img}{An \code{Image} object.}

\item{...}{Passed to \code{img$morph("gradient", ...)}.}
}
\value{
A new \code{Image}.
}
\description{
Compute the morphological gradient of an image
}
\examples{
\donttest{
img_path <- system.file("img", "flower.jpg", package = "Retina")
img <- Image$new(img_path)$convert_color("GRAY")
morph_gradient(img)$plot()
}
}
