% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology.R
\name{morph_open}
\alias{morph_open}
\title{Apply morphological opening to an image}
\usage{
morph_open(img, ...)
}
\arguments{
\item{img}{An \code{Image} object.}

\item{...}{Passed to \code{img$morph("open", ...)}.}
}
\value{
A new \code{Image}.
}
\description{
Apply morphological opening to an image
}
\examples{
\donttest{
img_path <- system.file("img", "flower.jpg", package = "Retina")
img <- Image$new(img_path)$convert_color("GRAY")
morph_open(img)$plot()
}
}
