% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\name{path_to_main_file_for_test_retina}
\alias{path_to_main_file_for_test_retina}
\title{path_to_main_file_for_test_retina
useful for access of the test retina datasets'}
\usage{
path_to_main_file_for_test_retina(string_identifier)
}
\arguments{
\item{string_identifier}{character e.g. "40oik5" that represents the inst/extdata identifier folder name.}
}
\value{
filepath to the test_retina main.r that matches the string_identifier
}
\description{
path_to_main_file_for_test_retina
useful for access of the test retina datasets'
}
