% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\name{plot_circle_radial_lines}
\alias{plot_circle_radial_lines}
\title{Draw circle radians about the center}
\usage{
plot_circle_radial_lines(number_of_circles, color_hex = "#66666650")
}
\arguments{
  \item{number_of_circles}{the number of radius lines to plot}
}
\description{
Draw N radius lines (circles)
}
\author{
Brian Cohn
}

