% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_degree_label_for_latitudes}
\alias{plot_degree_label_for_latitudes}
\title{Interpolate Input Data with Thin Plate Spline}
\usage{
plot_degree_label_for_latitudes(outer.radius, circle.rads)
}
\arguments{
\item{outer.radius}{the extent of the radius of the plot}

\item{circle.rads}{the number of radian circles that are drawn}

\item{minitics}{values referring to the circle}

\item{x}{input variable 1}

\item{y}{input variable 2}

\item{z}{response variable}

\item{lambda}{TPS parameter}

\item{polynomial_m}{TPS parameter}

\item{extrapolate}{true/false whether we should interpolate past the points, all the way to the eye equator}
}
\description{
Interpolation

Plot degree numbers
}
\author{
Brian Cohn

Brian Cohn
}
