% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\name{plot_falciform_process}
\alias{plot_falciform_process}
\title{Plot falciform process}
\usage{
plot_falciform_process(falciform_x, falciform_y)
}
\arguments{
  \item{falciform_x}{numeric vector of x coordinates}
  \item{falciform_y}{numeric vector of y coordinates}
}
\description{
smooths and plots the falciform process
}
\author{
Brian Cohn
}

