% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot_from_MAT}
\alias{plot_from_MAT}
\title{Plot from matrix}
\usage{
plot_from_MAT(z1, z2, MATRIX, contours = TRUE, legend = TRUE, axes = TRUE,
  points = TRUE, extrapolate = FALSE, col_breaks_source = 2,
  col_levels = 10, col = rev(colorRampPalette(brewer.pal(11,
  "PuOr"))(col_levels)), contour_breaks_source = 1,
  contour_levels = col_levels + 1, outer.radius = pi/2,
  circle.rads = pretty(c(0, outer.radius)), spatial_res = 128,
  single_point_overlay = 0, interp.type = 1, lambda = 0.001,
  xyrelief = 1, ...)
}
\arguments{
  \item{contours}{whether to plot contours.}

  \item{legend}{Color legend with tick marks}

  \item{axes}{Radial axes}

  \item{points}{whether to plot individual datapoints as
  X's}

  \item{extrapolate}{By default TRUE, will make plot a
  circle.}

  \item{col_breaks_source}{2 element vector with max and
  min}

  \item{col_levels}{number of color levels}

  \item{col}{colors to plot}

  \item{contour_breaks_source}{1 if data, 2 if calculated
  surface data}

  \item{contour_levels}{number of contour levels}

  \item{outer.radius}{size of plot}

  \item{circle.rads}{radius lines}

  \item{spatial_res}{Used to define a spatial_res by
  spatial_res plotting resolution.}

  \item{single_point_overlay}{Overlay "key" data point with
  square (0 = No, Other = number of pt)}

  \item{interp.type}{depreciated}

  \item{lambda}{lambda value for thin plate spline
  interpolation}

  \item{xyrelief}{scaling factor for interpolation matrix.}

  \item{z1}{Vector of density values (used to get the max
  and min)}

  \item{z2}{Vector of density values (used to get the max
  and min)}

  \item{MATRIX}{Matrix object containing densty values at
  each row and col}

  \item{...}{further arguments passed to or from other
  methods.}
}
\description{
Plot from matrix
}
\details{
Takes in a square matrix of azimuthal equidistant plot
projection points. Generates plot projection.
}

