% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_longitudinal_labels}
\alias{plot_longitudinal_labels}
\title{Plot longitudinal labels}
\usage{
plot_longitudinal_labels(axis.rad, outer_radius, r_label, l_label, degree)
}
\arguments{
\item{axis.rad}{axis radian}

\item{outer_radius}{numeric value of the outer radius limit}

\item{r_label}{label number}

\item{l_label}{label number}

\item{degree}{numeric, the degree of interest}
}
\description{
put a degree label at the ends of the endpoints for each longitude
}
\author{
Brian Cohn
}
