% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_original_xy_locations}
\alias{plot_original_xy_locations}
\title{Add scatter X's of XY to an existing plot}
\usage{
plot_original_xy_locations(x, y)
}
\arguments{
\item{x}{numeric vector of x coordinates}

\item{y}{numeric vector of y coordinates}
}
\description{
visualize xy points as X's
}
\author{
Brian Cohn
}
