% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_radial_spokes_and_labels}
\alias{plot_radial_spokes_and_labels}
\title{Plot radial axes}
\usage{
plot_radial_spokes_and_labels(outer_radius)
}
\arguments{
\item{outer_radius}{numeric value of the outer radius limit}
}
\description{
Put radial axes onto visualization
}
\author{
Brian Cohn
}
