% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\name{plot_rotation_optimize}
\alias{plot_rotation_optimize}
\title{Plot rotation optimization scatter plot with minima.}
\usage{
plot_rotation_optimize(rotation_df)
}
\arguments{
\item{rotation_df}{data.frame from spin optimization}
}
\description{
A base-graphics plot (rotation along the x axis,
Absolute Mean difference on the y axis) of the rotation optimization values, and a vertical line denoting the optimized rotation value.
}
