% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{pretty_list_not_including_max}
\alias{pretty_list_not_including_max}
\title{Create a pretty number list up to but not including the upper limit}
\usage{
pretty_list_not_including_max(lower_limit, upper_limit)
}
\arguments{
\item{lower_limit}{numeric lower bound}

\item{upper_limit}{numeric upper bound}
}
\value{
pretty_vec numeric vector
}
\description{
Remove the last element if it exceeds the upper limit
}
\author{
Brian Cohn
}
