% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reflect_across_vertical_line}
\alias{reflect_across_vertical_line}
\title{Reflect matrix across vertical axis}
\usage{
reflect_across_vertical_line(mat)
}
\arguments{
  \item{mat}{a matrix (doesn't have to be square)}
}
\value{
matrix reflected matrix
}
\description{
Reflect matrix across vertical axis
}
\author{
Brian Cohn \email{brian_cohn14@pitzer.edu} Lars Schmitz
}

