% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{reflect_IJ_y_values_for_falciform}
\alias{reflect_IJ_y_values_for_falciform}
\title{Reflect IJ Y values for the falciform process}
\usage{
reflect_IJ_y_values_for_falciform(falciform_xy_coordinates, xy_img_dimensions)
}
\arguments{
\item{falciform_xy_coordinates}{2 columns of coordinate (X,Y) data in ImageJ pixel format}

\item{xy_img_dimensions}{2-cal list i.e. c(400,200)}
}
\value{
falciform_xy_coordinates_transformed
}
\description{
ImageJ values don't always show up the right way, because Y values get larger and positive while moving down the image.
}
\author{
Brian Cohn
}
