% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\name{reorder_columns}
\alias{reorder_columns}
\title{Reorder columns by function}
\usage{
reorder_columns(X, FN)
}
\arguments{
  \item{X}{data.frame} \item{FN}{A function that takes in a
  vector and returns one value.}
}
\value{
X_new data.frame of columns ordered in decreasing value from
left to right according to each function(column)
}
\author{
Brian Cohn \email{brian_cohn14@pitzer.edu} Lars Schmitz
}

