% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Retina-package.R
\docType{package}
\name{Retina-package}
\alias{Retina}
\alias{Retina-package}
\title{Retina: Computer Vision Library for R}
\description{
An OpenCV-based computer vision library providing an idiomatic R interface to image and video processing. Images are managed as R6 objects backed by OpenCV's cv::Mat (CPU) or cv::UMat (GPU).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/swarm-lab/Retina}
  \item Report bugs at \url{https://github.com/swarm-lab/Retina/issues}
}

}
\author{
\strong{Maintainer}: Simon Garnier \email{garnier@njit.edu} (\href{https://orcid.org/0000-0002-3886-3974}{ORCID})

}
\keyword{internal}
