% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\docType{package}
\name{retina}
\alias{retina}
\alias{retina-package}
\title{retina.}
\description{
Copyright (c) 2013-2014 Brian Cohn and Lars Schmitz
}
\author{
Brian Cohn \email{brian_cohn14@pitzer.edu}, Lars Schmitz
}
