% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\name{retinal_phi0}
\alias{retinal_phi0}
\title{Rim Latitude Estimation (Ellipsoidal assumption)}
\usage{
retinal_phi0(ED, AL)
}
\arguments{
\item{ED}{Eye Diameter of the eye sample, measured at the widest points of the eye.}

\item{AL}{Axial Length of the eye sample, measured at the longest measure of the eye from anterior to posterior.}
}
\value{
Retinal latitude in degrees
}
\description{
Rim Latitude Estimation (Ellipsoidal assumption)
}
\details{
Assumes an ellipsoidal eye
}
\author{
Brian Cohn \email{brian_cohn14@pitzer.edu}, Lars Schmitz
}
