% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/retina-package.r
\name{retinaplot}
\alias{retinaplot}
\title{Retina Plot}
\usage{
retinaplot(retina_object, spatial_res = 1000, rotation = 0,
  inner_eye_view = TRUE, return_fit = FALSE, ...)
}
\arguments{
  \item{retina_object}{A list containing an element
  \code{azimuthal_data.datapoints} with \code{x,y,z}
  datapoints. File must also include
  \code{azimuthal_data.falciform}.}

  \item{spatial_res}{define the number of pixels
  (resolution) the plot will be}

  \item{rotation}{degrees to rotate CCW (int or floating
  point)}

  \item{inner_eye_view}{boolean, default is TRUE. If set to
  false, the plotted view of the retina will have the
  viewpoint within the skull looking at the rear of the
  eye. inner_eye_view has the same view as the traditional
  wholemount.}

  \item{return_fit}{logical, whether or not to return the
  interpolation fit data.}

  \item{...}{further arguments passed to or from other
  methods.}
}
\value{
Base-R polar plot
}
\description{
\code{retinaplot} Generates an Azimuthal Equidistant plot
projection of a retina object.
}
\author{
Brian Cohn \email{brian_cohn14@pitzer.edu}, Lars Schmitz
}

