% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{RMat}
\alias{RMat}
\title{Internal equation for axis markup}
\usage{
RMat(radians)
}
\arguments{
\item{radians}{vector of radians}
}
\value{
RMat trigonometric positions
}
\description{
define locations for radians with trigonometry
}
\author{
Brian Cohn
}
