% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\name{run_all_demos}
\alias{run_all_demos}
\title{Run all demos}
\usage{
run_all_demos(path_to_demo_folder = file.path(.libPaths(), "retina/demo"))
}
\arguments{
\item{path_to_demo_folder}{include the trailing /}
}
\description{
Runs all demos in quick succession
}
\author{
Brian Cohn \email{brian.cohn@usc.edu}, Lars Schmitz
}
