% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\name{set_contour_breaks_based_on_requested_source}
\alias{set_contour_breaks_based_on_requested_source}
\title{Set Contour Breaks Based on Requested Source}
\usage{
set_contour_breaks_based_on_requested_source(contour_breaks_source, z,
  contour_levels, Mat)
}
\arguments{
  \item{contour_breaks_source}{See fit_plot_azimuthal}
  \item{z}{See fit_plot_azimuthal} \item{contour_levels}{See
  fit_plot_azimuthal} \item{Mat}{See fit_plot_azimuthal}
}
\description{
This function will make a set of contour topography lines.
Code from http://stackoverflow.com/questions/10856882/r-
interpolated-polar-contour-plot was highly modified to meet
retinal plotting funtionality.
}

