% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R, R/poly_smoothing.R
\name{spline_poly}
\alias{spline_poly}
\alias{spline_poly}
\title{Polygon Spline Fit}
\usage{
spline_poly(xy_dataframe, vertices, k = 3, ...)

spline_poly(xy_dataframe, vertices, k = 3, ...)
}
\arguments{
\item{xy_dataframe}{xy dataframe coordinates, in two columns.}

\item{vertices}{Number of spline vertices to create.}

\item{k}{Wraps K vertices around each end. n >=k}

\item{...}{further arguments passed to or from other methods.}

\item{xy}{vertices in dataframe with x and y columns, in order (not all are used).}

\item{k}{number of points to wrap around ends to obtain smooth spline}

\item{...}{further arguments to spline}
}
\value{
Coords More finely placed vertices for the polygon.

xy_dataframe_splined xy dataframe coordinates, in two columns, with new points from spline.
}
\description{
Useful for making the falciform process look more smooth and refined. This is purely aesthetic.

The rows of 'xy_dataframe' give coordinates of the boundary
vertices, in order.  'vertices' is the number of spline vertices to create.
(Not all are used: some are clipped from the ends.)  'k' is the number of
points to wrap around the ends to obtain a smooth periodic spline.  reference:
http://gis.stackexchange.com/questions/24827/how-to-smooth-the-polygons-in-a-contour-map
Returns an array of points.
}
\details{
enhance the resolution of a polygon vertices dataframe by creating a spline along each vertex.
}
\references{
http://gis.stackexchange.com/questions/24827/how-to-smooth-the-polygons-in-a-contour-map
}
\author{
Brian Cohn \email{brian.cohn@usc.edu}, Lars Schmitz

Brian Cohn \email{brian.cohn@usc.edu}, Lars Schmitz
}
