% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels.R
\name{split_channels}
\alias{split_channels}
\title{Split an image into its individual channels}
\usage{
split_channels(img)
}
\arguments{
\item{img}{An \code{Image} object.}
}
\value{
A named list of single-channel \code{Image} objects. Names are
  derived from the colorspace (e.g. \code{c("B", "G", "R")} for BGR).
  Unknown colorspaces use \code{"ch1"}, \code{"ch2"}, etc.
}
\description{
Split an image into its individual channels
}
\examples{
\donttest{
img <- Image$new(system.file("img", "flower.jpg", package = "Retina"))
channels <- split_channels(img)
channels$B$plot()
}
}
