% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\name{ssite_merge}
\alias{ssite_merge}
\title{Merge Sampling Site location and Counting Frame counts.}
\usage{
ssite_merge(location, counts, ...)
}
\arguments{
\item{location}{dataframe with a column for samplingsite, x and y.}

\item{counts}{dataframe with a column for samplingsite and count.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
xycount Dataframe of (x,y, count)
}
\description{
X represents the row index on the sampling grid, and Y represents the column index. Numbers start at 1.
}
\seealso{
Other munge: \code{\link{import_xyz}}
}
\author{
Brian Cohn \email{brian.cohn@usc.edu}, Lars Schmitz
}
