% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\name{str_eval}
\alias{str_eval}
\title{String evaluation via eval
useful for load_function_and_run}
\usage{
str_eval(x)
}
\arguments{
\item{x}{a string that has the command to run. e.g. "main_40oik5("path/of/interest")"}
}
\description{
String evaluation via eval
useful for load_function_and_run
}
