% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{tear_markup_plot}
\alias{tear_markup_plot}
\title{Tear Markup Plot}
\usage{
tear_markup_plot(path_to_retina_data_folder)
}
\arguments{
\item{path_to_retina_data_folder}{The path to the folder that contains the outline.roi file.}
}
\value{
roi_coordinates XY vals of pixel coordinates of the outline points, in order.
}
\description{
Shows a visualization of the outline, as well as the location of each datapoint so the user can identify where the tears are.
This is especially useful when retistruct is nonfunctioning.
}
\author{
Brian Cohn
}
