% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{traditional_flatmount_plot}
\alias{traditional_flatmount_plot}
\title{Traditional Flatmount Plot TODO: implement.}
\usage{
traditional_flatmount_plot(pixel_xy_positions, z_vector, xy_outline_points)
}
\arguments{
\item{pixel_xy_positions}{2 columns of coordinate (X,Y) data in ImageJ pixel format}

\item{z_vector}{numerical vector of density values for each of the coordinates}

\item{xy_outline_points}{2 columns of coordinate(X,Y) data in ImageJ pixel format for the outline coordinates.}
}
\value{
fit the Fit object generated by Tps
}
\description{
Shows a visualization of the outline, as well as the location of each datapoint, with fit plot.
}
\author{
Brian Cohn
}
