% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retistruct_processing_functions.R
\name{update_outline_object_tears}
\alias{update_outline_object_tears}
\title{Set up tear matrix within the annotated outline object}
\usage{
update_outline_object_tears(outline_object, tear_coordinates_dataframe)
}
\arguments{
\item{outline_object}{See ?retistruct::labelTearPoints for documentation}

\item{tear_coordinates_dataframe}{dataframe with $V0, $VB, and $VF columns as produced by compose_tear_triplets_dataframe}
}
\value{
annotated_outline See ?retistruct::labelTearPoints for documentation
}
\description{
assembles the outline object with tear data at $V0, $VB, and $VF
}
\author{
Brian Cohn \email{brian.cohn@usc.edu}
}
