% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{vector_retina_composite}
\alias{vector_retina_composite}
\title{Generate a retinal composite}
\usage{
vector_retina_composite(retina_list, spin_spatial_res = 128,
  plot_spatial_res = 512, theta_interval = 10, plot = FALSE,
  rotation_degree_list, ...)
}
\arguments{
\item{retina_list}{A list containing retina objects, with names(retina_list) all defined.}

\item{spin_spatial_res}{Resolution passed to the spin-optimization.}

\item{plot_spatial_res}{Resolution passed to the final plotter}

\item{theta_interval}{number of degrees to traverse in spin-optimization}

\item{plot}{Logical, by default false, but when true it will output the plot where the first element in the list is fixed.}

\item{...}{Arguments passed from other functions. You can access the thin plate spline interpolator here.}
}
\value{
composite_mat Matrix of the composite retina, where the first element of the retina list is fixed.
}
\author{
Brian Cohn \email{brian_cohn14@pitzer.edu} and Lars Schmitz
}
