% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{write_labels_at_endpoint_locations}
\alias{write_labels_at_endpoint_locations}
\title{Write labels at endpoint locations}
\usage{
write_labels_at_endpoint_locations(r_label, l_label, degree, endpoints)
}
\arguments{
\item{r_label}{string of right label}

\item{l_label}{string of left label}

\item{degree}{the degree that is being placed in}

\item{endpoints}{vector of 4 numerics, x,y and x',y' defining the line segment}
}
\description{
put labels around the circle at each of the lines
}
\author{
Brian Cohn
}
