% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\name{composite_map}
\alias{composite_map}
\title{Make a Composite Map}
\usage{
composite_map(map1, map2, rotation = TRUE, spatial_res,
  plot_rotation = FALSE, projection_type = "azequidistant", ...)
}
\arguments{
\item{map1}{Fixed retina object}

\item{map2}{retina object subject to rotation.}

\item{rotation}{Boolean, whether or not spin optimization is used.}

\item{spatial_res}{Default is 16. A (spatial_res*spatial_res) matrix will be created for spin optimization comparisons and compose the average.}

\item{plot_rotation}{Boolean, whether or not spin optimization plot is shown.}

\item{projection_type}{A mapproject mapproj projection type}

\item{...}{further arguments passed to or from other methods.}
}
\value{
plot Composite Plot data
}
\description{
Combines two retinal cell density maps
}
\seealso{
Other species_average: \code{\link{map_composites}},
  \code{\link{optimal_rotation}},
  \code{\link{rotation_optimize}}
}
\author{
Brian Cohn \email{brian.cohn@usc.edu}, Lars Schmitz
}
