% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{draw_latitude_markings}
\alias{draw_latitude_markings}
\title{Draw Latitude Markings}
\usage{
draw_latitude_markings(radius_vals, outer_radius)
}
\arguments{
\item{radius_vals}{vector of radius values (for latitudes)}

\item{outer_radius}{see fitplotazimuthal}
}
\description{
plots radial lines, degree label for latitudes, and plots radial spokes with labels
}
\author{
Brian Cohn
}
