% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dss_retistruct_processing}
\alias{dss_retistruct_processing}
\title{Retistruct Wrapper}
\usage{
dss_retistruct_processing(path)
}
\arguments{
  \item{path}{string path to the folder with retinal data.}
}
\value{
rad Retinal data in radian units.
}
\description{
Reads in the folder, reconstructs, and returns the retinal
object.
}
\author{
Brian Cohn \email{brian_cohn14@pitzer.edu}
}
\references{
Sterratt et. al. 2013
}

