% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\name{fit_error_histogram}
\alias{fit_error_histogram}
\title{Single Tps Fit Error Plot}
\usage{
fit_error_histogram(x, main = NULL, ...)
}
\arguments{
\item{x}{the fit object}

\item{main}{title of the plot, NULL by default}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Error histogram in base R
}
\author{
Brian Cohn \email{brian_cohn14@pitzer.edu}, Lars Schmitz
}
