% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fit_error_histogram}
\alias{fit_error_histogram}
\title{Single Tps Fit Error Plot}
\usage{
fit_error_histogram(x, main = NULL, ...)
}
\arguments{
  \item{x}{the fit object}

  \item{main}{title of the plot, NULL by default}

  \item{...}{further arguments passed to or from other
  methods.}
}
\value{
Error histogram in base R
}
\description{
Single Tps Fit Error Plot
}
\author{
Brian Cohn \email{brian_cohn14@pitzer.edu}, Lars Schmitz
}

