% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\name{map_composites}
\alias{map_composites}
\title{Composite two Retinal Map Matrices}
\usage{
map_composites(MAT1, MAT2, col_levels = 5, showplots = FALSE, ...)
}
\arguments{
\item{MAT1}{Matrix of RGC densities}

\item{MAT2}{Matrix of RGC densities}

\item{col_levels}{Number of levels to plot (if showing plot.)}

\item{showplots}{False by default.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
MAT_composite Square matrix: Composite Map (adjusted to max,min of each map, with new peak equal to the mean of map1 and map2's peaks)
}
\description{
Combines two density matrices of pre-smoothed data.
}
\author{
Brian Cohn \email{brian.cohn@usc.edu}, Lars Schmitz
}
