% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\name{nullify_vals_outside_the_circle}
\alias{nullify_vals_outside_the_circle}
\title{Remove points that are outside of the plotting circle}
\usage{
nullify_vals_outside_the_circle(minitics, spatial_res, heatmap_matrix,
  outer.radius)
}
\arguments{
  \item{minitics}{Spherical limit info} \item{spatial_res}
  {spatial width in pixels of the plotted image}
  \item{outer.radius}{max value of the radius} \item{Mat}
  {Matrix of predicted points on the set grid}
  \item{falciform_y}{numeric vector of y coordinates}
}
\description{
We do not need points plotted in the corners of the plotted
circle.
}
\author{
Brian Cohn
}

