% Generated by roxygen2 (4.0.0): do not edit by hand
\name{optimal_rotation}
\alias{optimal_rotation}
\title{Optimal rotation (for min_err) for rotation error matrix
Computes the argmax of the min error, returning both the error and its theta rotation value in degrees.}
\usage{
optimal_rotation(df_spin, quiet = TRUE)
}
\arguments{
  \item{df_spin}{Two column matrix of theta (rotation in
  degrees) at [,1] and error at [,2].}

  \item{quiet}{logical. when false, it will print what the
  absolute mean difference is between maps.}
}
\value{
optimal_degree The optimal (minimum) degree and its error
value.
}
\description{
Optimal rotation (for min_err) for rotation error matrix
Computes the argmax of the min error, returning both the
error and its theta rotation value in degrees.
}
\author{
Brian Cohn \email{brian_cohn14@pitzer.edu}, Lars Schmitz
}
\seealso{
Other species_average: \code{\link{composite_map}};
\code{\link{rotation_optimize}}
}

