% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_circle_radial_lines}
\alias{plot_circle_radial_lines}
\title{Draw circle radians about the center}
\usage{
plot_circle_radial_lines(radius_vals, color_hex = "#66666650")
}
\arguments{
\item{radius_vals}{vector of radius values where the circles will be drawn}
}
\description{
Draw N radius lines (circles)
}
\author{
Brian Cohn
}
