% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_degree_label_for_latitudes}
\alias{plot_degree_label_for_latitudes}
\title{Plot the degree labels for latitudes}
\usage{
plot_degree_label_for_latitudes(outer_radius, circle.rads)
}
\arguments{
\item{outer_radius}{the extent of the radius of the plot}

\item{circle.rads}{the number of radian circles that are drawn}
}
\description{
Plot degree numbers
}
\author{
Brian Cohn
}
