% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_longitudinal_lines}
\alias{plot_longitudinal_lines}
\title{Plot longitudinal spoke lines}
\usage{
plot_longitudinal_lines(axis_radian, outer_radius)
}
\arguments{
\item{axis_radian}{radian}

\item{outer_radius}{numeric value of the outer radius limit}
}
\description{
put lines across the plotting circle
}
\author{
Brian Cohn
}
