% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_original_xy_locations}
\alias{plot_original_xy_locations}
\title{plot points of XY}
\usage{
plot_original_xy_locations(x, y)
}
\arguments{
\item{falciform_x}{numeric vector of x coordinates}

\item{falciform_y}{numeric vector of y coordinates}
}
\description{
visualize xy points
}
\author{
Brian Cohn
}
