% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{reef_fish}
\alias{reef_fish}
\title{Reef Fish}
\format{A list (in retina object form)}
\usage{
data(reef_fish)
}
\description{
A dataset containing three retinal objects, computed using
the retina package: Pmol 753,Pmol 752 and Ntae 381.
}
\author{
Brian Cohn \email{brian_cohn14@pitzer.edu}, Lars Schmitz
}
\keyword{datasets}

