% Generated by roxygen2 (4.0.2): do not edit by hand
\name{retinal_arclen}
\alias{retinal_arclen}
\title{' Retinal Perimeter Estimation (spherical assumption)}
\usage{
retinal_arclen(ED)
}
\arguments{
  \item{ED}{Eye Diameter of the eye sample, measured at the
  widest point of the eye.}
}
\value{
Retinal Arclength in millimeters
}
\description{
Calculates the retinal perimeter distance in same units as
input, with the assumption that the retinal shape is
hemispherical.
}
\details{
Assumes a spherical eye, where eye diameter approximates
twice the sphere's radius.
}
\author{
Brian Cohn \email{brian_cohn14@pitzer.edu}, Lars Schmitz
}

