% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/retina-package.r
\name{sphere_visualize}
\alias{sphere_visualize}
\title{Spherical Plot visualization
Uses retistruct to create lat/lon coordinates}
\usage{
sphere_visualize(trimmed_data)
}
\arguments{
  \item{trimmed_data}{Three-column dataset with datapoint
  locations in latitude/longitude degree format.}
}
\value{
Spherical visualization
}
\description{
Spherical Plot visualization Uses retistruct to create
lat/lon coordinates
}
\author{
Brian Cohn \email{brian_cohn14@pitzer.edu}, Lars Schmitz
}

