% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retina-package.r
\name{spherical_coords}
\alias{spherical_coords}
\title{ImageJ and Retistruction of Input Retinal Data from Flattened State}
\usage{
spherical_coords(path, height, width, IJ_limits, falciform = TRUE)
}
\arguments{
\item{path}{Directory which contains the retistruct files.}

\item{height}{Height in microns of the stereologic counting frame}

\item{width}{Width in microns of the stereologic counting frame}

\item{IJ_limits}{data.frame with min, max and delta values for both X and Y with respect to the contour image in ImageJ coordiantes. For setting the calibration of the sampling sites to the sphere.}

\item{falciform}{(boolean) True by default, meaning there is a file called falc.txt within the path.}
}
\value{
data.frame with phi(latitude), lambda(longitude) and Z (cells per square millimeter).
}
\description{
Converts flat retina to a spherical representation of the data. Spherical coords path must contain:  outline.roi  Datapoints in barycentric coordinates. P.csv  Datapoints on reconstructed sphere in cartesian coordinates. T.csv  Datapoints on reconstructed sphere in spherical coordinates.  xyz.csv  Cell density site locations with density measurements. Columns of x(integer column starting at 1 at the left) , y (integer column starting at 1 on the bottom) and z (counts)
}
\references{
https://r-forge.r-project.org/scm/?group_id=1436
}
\author{
Brian Cohn \email{brian.cohn@usc.edu}, Lars Schmitz
}
