% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{spline.poly}
\alias{spline.poly}
\title{Polygon Spline Fit}
\usage{
spline.poly(xy, vertices, k = 3, ...)
}
\arguments{
\item{xy}{vertices in dataframe with x and y columns, in order (not all are used).}

\item{vertices}{Number of spline vertices to create.}

\item{k}{Wraps K vertices around each end. n >=k}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Coords More finely placed vertices for the polygon.
}
\description{
Useful for making the faliform process look more smooth and refined. This is purely aesthetic.
}
\details{
enhance the resolution of a polygon verticies dataframe by creating a spline along each vertex.
}
\references{
http://gis.stackexchange.com/questions/24827/how-to-smooth-the-polygons-in-a-contour-map
}
\author{
Brian Cohn \email{brian.cohn@usc.edu}, Lars Schmitz
}
